backend/requirements.txt
# Core Framework
fastapi==0.120.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator>=2.1.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6

# LangGraph & AI
langgraph==0.2.34
langchain==0.2.16
langchain-openai==0.1.25
openai==1.99.9

# HTTP & API
httpx==0.26.0
requests==2.31.0

# Build & Review Tooling
playwright
pillow
opencv-python
beautifulsoup4
lxml

# Email Service (Resend)
resend==0.7.0

# AWS S3
boto3==1.34.34
botocore==1.34.34

# Redis & Rate Limiting
redis==5.0.1
slowapi==0.1.9

# Error Tracking
sentry-sdk[fastapi]==1.40.0

# File Validation
python-magic==0.4.27

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# GraphQL
strawberry-graphql[fastapi]==0.216.0

# Analytics
pandas==2.1.4

# Templates
jinja2==3.1.2

# Utilities
python-dotenv==1.0.0
